+++ Source/Krypton Components/Krypton.Docking/Elements Impl/KryptonDockingDockspace.cs
@@ -1,12 +1,12 @@
 ﻿#region BSD License
 /*
- * 
+ *
  * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
  *  © Component Factory Pty Ltd, 2006 - 2016, (Version 4.5.0.0) All rights reserved.
  *
  *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
- *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2017 - 2023. All rights reserved. 
- *  
+ *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), tobitege et al. 2017 - 2026. All rights reserved.
+ *
  */
 #endregion
 
@@ -418,8 +418,6 @@ namespace Krypton.Docking
             if (DockspaceControl.PageCount == 0)
             {
                 DockspaceControl.Dispose();
-                // TODO: Is this safe ?. It means that whatever uses this afterwards could be accessing null things !
-                SpaceControl = null;
             }
         }
         #endregion
@@ -427,7 +425,12 @@ namespace Krypton.Docking
         #region Implementation
         private void OnDockspaceCellVisibleCountChanged(object sender, EventArgs e)
         {
-            if (DockspaceControl.CellVisibleCount == 0)
+            if (!(sender is KryptonDockspace dockspace))
+            {
+                return;
+            }
+
+            if (dockspace.CellVisibleCount == 0)
             {
                 if (_cacheCellVisibleCount > 0)
                 {
@@ -447,10 +450,15 @@ namespace Krypton.Docking
 
         private void OnDockspaceCellCountChanged(object sender, EventArgs e)
         {
-            // When all the cells (and so pages) have been removed we kill ourself
-            if (DockspaceControl.CellCount == 0)
+            if (!(sender is KryptonDockspace dockspace))
             {
-                DockspaceControl.Dispose();
+                return;
+            }
+
+            // When all the cells (and so pages) have been removed we kill ourself
+            if (dockspace.CellCount == 0)
+            {
+                dockspace.Dispose();
             }
         }